requirement2.txt
ultralytics==8.0.227
Flask-SQLAlchemy==3.1.1
Flask==3.0.0